+++ packages/abstarct-connector/src/index.ts
@@ -1,4 +1,11 @@
-import { Connector, DefaultConnectionPayload } from '@web3-wallets-kit/types';
+import {
+  ChainIdCallback,
+  ConnectCallback,
+  Connector,
+  DefaultConnectionPayload,
+  DisconnectCallback,
+  SubscribedObject,
+} from '@web3-wallets-kit/types';
 import { getAccount, getChainId, SendingInterface } from '@web3-wallets-kit/utils';
 
 export abstract class AbstractConnector<P extends DefaultConnectionPayload>
@@ -6,7 +13,7 @@ export abstract class AbstractConnector<P extends DefaultConnectionPayload>
   protected payload: P | null = null;
   private sendingInterface: SendingInterface = 'EIP 1193';
 
-  public abstract async connect(): Promise<P>;
+  public abstract connect(): Promise<P>;
 
   public async disconnect() {
     this.payload = null;
@@ -44,5 +51,50 @@ export abstract class AbstractConnector<P extends DefaultConnectionPayload>
 
   public getConnectionPayload() {
     return this.payload;
+  }
+
+  public subscribeConnectAccount(callback: ConnectCallback): SubscribedObject {
+    const convertedCallback = (accounts: string[]) => callback(accounts[0]);
+
+    this.payload?.provider.on && this.payload.provider.on('accountsChanged', convertedCallback);
+
+    return {
+      unsubscribe: () => {
+        this.payload?.provider.removeListener &&
+          this.payload.provider.removeListener('accountsChanged', convertedCallback);
+      },
+    };
+  }
+
+  public subscribeChainId(callback: ChainIdCallback): SubscribedObject {
+    const convertedCallback = (chainId: number | string) => {
+      const convertedChainId = typeof chainId === 'string' ? parseInt(chainId, 16) : chainId;
+
+      if (Number.isNaN(convertedChainId)) {
+        throw new Error('ChainId is incorrect');
+      } else {
+        callback(convertedChainId);
+      }
+    };
+
+    this.payload?.provider.on && this.payload.provider.on('chainChanged', convertedCallback);
+
+    return {
+      unsubscribe: () => {
+        this.payload?.provider.removeListener &&
+          this.payload.provider.removeListener('chainChanged', convertedCallback);
+      },
+    };
+  }
+
+  public subscribeDisconnect(callback: DisconnectCallback): SubscribedObject {
+    this.payload?.provider.on && this.payload.provider.on('disconnect', callback);
+
+    return {
+      unsubscribe: () => {
+        this.payload?.provider.removeListener &&
+          this.payload.provider.removeListener('disconnect', callback);
+      },
+    };
   }
 }