+++ packages/portis-connector/src/index.ts
@@ -1,6 +1,12 @@
 /* eslint-disable import/no-duplicates */
 import { AbstractConnector } from '@web3-wallets-kit/abstract-connector';
-import { DefaultConnectionPayload, Provider } from '@web3-wallets-kit/types';
+import {
+  ConnectCallback,
+  DefaultConnectionPayload,
+  DisconnectCallback,
+  Provider,
+  SubscribedObject,
+} from '@web3-wallets-kit/types';
 import type PortisClass from '@portis/web3';
 import type { IOptions, INetwork } from '@portis/web3';
 
@@ -50,5 +56,21 @@ export class PortisConnector extends AbstractConnector<PortisConnectionPayload> 
       this.payload.portis.logout();
     }
     super.disconnect();
+  }
+
+  public subscribeConnectAccount(callback: ConnectCallback): SubscribedObject {
+    this.payload?.portis.onActiveWalletChanged(callback);
+
+    return {
+      unsubscribe: () => this.payload?.portis.onActiveWalletChanged(() => {}),
+    };
+  }
+
+  public subscribeDisconnect(callback: DisconnectCallback): SubscribedObject {
+    this.payload?.portis.onLogout(callback);
+
+    return {
+      unsubscribe: () => this.payload?.portis.onLogout(() => {}),
+    };
   }
 }