+++ packages/connect-wallet-connector/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.8 (2021-05-14)
+
+**Note:** Version bump only for package @web3-wallets-kit/connect-wallet-connector
+
+
+
+
+
+## 0.1.7 (2021-05-06)
+
+**Note:** Version bump only for package @web3-wallets-kit/connect-wallet-connector
+
+
+
+
+
+## 0.1.6 (2021-04-20)
+
+**Note:** Version bump only for package @web3-wallets-kit/connect-wallet-connector
+
+
+
+
+
 ## 0.1.5 (2021-04-12)
 
 **Note:** Version bump only for package @web3-wallets-kit/connect-wallet-connector